validacao para excluir comentario

+++ JSP/avaliacao_jsp/src/main/webapp/excluirComentario.jsp
@@ -11,19 +11,31 @@
 <body>
 	<%
 	
-		int codigoComentario = Integer.parseInt(request.getParameter("codigoComentario"));
-		int codigo = Integer.parseInt(request.getParameter("codigo"));
-		
-		Conexao c = new Conexao();
-		
-		String deleteCometario = "DELETE FROM comentario WHERE codigo = ?";
-		
-		PreparedStatement pstmt = c.efetuarConexao().prepareStatement(deleteCometario);
-		pstmt.setInt(1, codigoComentario);
-		
-		pstmt.execute();
-		
-		response.sendRedirect("detalhePublicacao.jsp?codigo="+codigo);
+		String email=(String)session.getAttribute("email");
+	
+		if(email != null) {
+			try {
+				int codigoComentario = Integer.parseInt(request.getParameter("codigoComentario"));
+				int codigo = Integer.parseInt(request.getParameter("codigo"));
+				
+				Conexao c = new Conexao();
+				
+				String deleteCometario = "DELETE FROM comentario WHERE codigo = ?";
+				
+				PreparedStatement pstmt = c.efetuarConexao().prepareStatement(deleteCometario);
+				pstmt.setInt(1, codigoComentario);
+				
+				pstmt.execute();
+				
+				response.sendRedirect("detalhePublicacao.jsp?codigo="+codigo);
+			} catch(Exception e) {
+				response.sendRedirect("index.jsp");
+			}
+		} else {
+			response.sendRedirect("acessarConta.jsp");
+		}
+	
+
 	%>
 </body>
 </html>
